refactor(profile): modularize components and split styles
- 將 Profile 元件內部結構進行拆分，提升模組化程度
- 拆分樣式至對應的元件檔案中，減少耦合
- 新增獨立的子元件，例如 Address, Contact, Disable 等
